Fedora: Use older images for ppc64le and s390x
On s390x we have this problem that I could not resolve with CONFIG_SHELL=/bin/bash
and M$=/usr/bib/m4. So, use an older version of Fedora.

/usr/bin/autoconf: This script requires a shell more modern than all
/usr/bin/autoconf: the shells that I found on your system.
/usr/bin/autoconf: Please tell bug-autoconf@gnu.org about your system,
/usr/bin/autoconf: including any error possibly output before this
/usr/bin/autoconf: message. Then install a modern shell, or manually run
/usr/bin/autoconf: the script under such a shell if you do have one.
autoreconf: /usr/bin/autoconf failed with exit status: 1

On ppc64le no Fedora image for this architecture seems to be available:

docker: no matching manifest for linux/ppc64le in the manifest list entries.

Signed-off-by: Stefan Berger <stefanb@linux.ibm.com>